2.4 Рефакторинг
