+++ src/const.js
@@ -1,0 +1,41 @@
+export const EVENT_TYPES = [
+  `Taxi`,
+  `Bus`,
+  `Train`,
+  `Ship`,
+  `Transport`,
+  `Drive`,
+  `Flight`,
+  `Check-in`,
+  `Sightseeing`,
+  `Restaurant`
+];
+
+export const DESTINATIONS = [
+  `Amsterdam`,
+  `Chamonix`,
+  `Geneva`
+];
+
+export const EXTRA_OPTIONS = [
+  {
+    title: `Add luggage`,
+    cost: 10,
+  },
+  {
+    title: `Switch to comfort class`,
+    cost: 20,
+  },
+  {
+    title: `Add meal`,
+    cost: 30,
+  },
+  {
+    title: `Choose seats`,
+    cost: 40,
+  },
+  {
+    title: `Travel by train`,
+    cost: 50,
+  },
+];
