+++ src/utils/common.js
@@ -1,5 +1,7 @@
 // Функция из интернета по генерации случайного числа из диапазона
 // Источник - https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_random
+import dayjs from "dayjs";
+
 export const getRandomInteger = (a = 0, b = 1) => {
   const lower = Math.ceil(Math.min(a, b));
   const upper = Math.floor(Math.max(a, b));
@@ -18,4 +20,9 @@ export const arrayShuffle = (array) => {
   }
   return array;
 };
+
+export const getFormatedDate = (date, format) => {
+  return date !== null ? dayjs(date).format(format) : ``;
+};
+
